feat: modify width
